refactor: 헥사고날 아키텍쳐 적용
