+++ dongmanee/src/main/java/com/dongmanee/domain/member/controller/port/MemberControllerUniversityService.java
@@ -1,12 +1,9 @@
-package com.dongmanee.domain.university.service;
+package com.dongmanee.domain.member.controller.port;
 
-import java.util.List;
 import java.util.Optional;
 
 import com.dongmanee.domain.university.domain.University;
 
-public interface UniversityService {
-	List<University> findAll();
-
+public interface MemberControllerUniversityService {
 	Optional<University> findById(Long universityId);
 }
