+++ dongmanee/src/main/java/com/dongmanee/domain/university/controller/UniversityController.java
@@ -5,10 +5,10 @@ import java.util.List;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
 
+import com.dongmanee.domain.university.controller.mapper.UniversityMapper;
+import com.dongmanee.domain.university.controller.port.UniversityControllerUniversityService;
 import com.dongmanee.domain.university.domain.University;
 import com.dongmanee.domain.university.dto.response.ResponseUniversity;
-import com.dongmanee.domain.university.mapper.UniversityMapper;
-import com.dongmanee.domain.university.service.UniversityService;
 import com.dongmanee.global.utils.ApiResponse;
 
 import lombok.RequiredArgsConstructor;
@@ -16,12 +16,12 @@ import lombok.RequiredArgsConstructor;
 @RestController
 @RequiredArgsConstructor
 public class UniversityController {
-	private final UniversityService universityService;
+	private final UniversityControllerUniversityService universityControllerUniversityService;
 	private final UniversityMapper universityMapper;
 
 	@GetMapping("/universities")
 	public ApiResponse<?> getAllUniversityList() {
-		List<University> universityList = universityService.findAll();
+		List<University> universityList = universityControllerUniversityService.findAll();
 		List<ResponseUniversity> responseUniversityList = universityList.stream()
 			.map(universityMapper::toResponseUniversity)
 			.toList();
