+++ dongmanee/src/main/java/com/dongmanee/domain/university/controller/port/UniversityControllerUniversityService.java
@@ -1,0 +1,9 @@
+package com.dongmanee.domain.university.controller.port;
+
+import java.util.List;
+
+import com.dongmanee.domain.university.domain.University;
+
+public interface UniversityControllerUniversityService {
+	List<University> findAll();
+}
